join all plaftorm files in one, no more externs
